Matchmaker: First pass complete of matchmaker
Matchmaker is now a walking skeleton. Arbitrary lists of players and requested team counts
can be provided and matches will be provided such that teams are equal in headcount and
the remainder returned as well.

Full completion of the acceptance criteria will require implementing alternative assignment
algorithms, but this is enough that we can start working on other parts of the solution.